+++ CHANGELOG.md
@@ -6,10 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+
+## [1.0.0-beta9] - 2025-08-14
 ### Added
 - default concurrency constructor parameter for TaskCommandDeps
 - container dependency is now optional
-
 
 ## [1.0.0-beta8] - 2025-08-14
 ### Fixed
@@ -78,7 +81,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2024-08-20
 - After months of development, the first release of Multitron is here!
 
-[Unreleased]: https://github.com/riki137/multitron/compare/1.0.0-beta8...master
+[Unreleased]: https://github.com/riki137/multitron/compare/1.0.0-beta9...master
+[1.0.0-beta9]: https://github.com/riki137/multitron/compare/1.0.0-beta8...1.0.0-beta9
 [1.0.0-beta8]: https://github.com/riki137/multitron/compare/1.0.0-beta7...1.0.0-beta8
 [1.0.0-beta7]: https://github.com/riki137/multitron/compare/1.0.0-beta6...1.0.0-beta7
 [1.0.0-beta6]: https://github.com/riki137/multitron/compare/1.0.0-beta5...1.0.0-beta6